feat(rds): Made suggested changes
